requirements.txt
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
blurhash-python==1.2.2
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
cloudinary==1.38.0
constantly==23.10.4
cryptography==42.0.1
daphne==4.0.0
defusedxml==0.7.1
dj-rest-auth==5.0.2
Django==5.0.1
django-allauth==0.60.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-dotenv==1.4.2
django-phonenumber-field==7.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
http_ece==1.2.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
oauthlib==3.2.2
packaging==23.2
phonenumberslite==8.13.29
pillow==10.2.0
psycopg==3.1.17
psycopg-binary==3.1.17
psycopg-pool==3.2.1
py-vapid==1.9.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==24.0.0
python3-openid==3.2.0
pytz==2023.3.post1
pywebpush==1.14.1
requests==2.31.0
requests-oauthlib==1.3.1
service-identity==24.1.0
setuptools==69.0.3
six==1.16.0
sqlparse==0.4.4
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
zope.interface==6.1
